rename table to base-table
